made everything more like a class
